requirements.txt
fastapi
uvicorn[standard]
pandas
numpy
scikit-learn
tensorflow
matplotlib
python-multipart
fpdf2 # For PDF generation